git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/new_version_support@2075 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/NPCs.cpp
@@ -101,13 +101,13 @@ void NPCs::handleNPCIn(Player *player, PacketReader &packet) {
 			npc->end();
 		}
 		else {
-			npc->setSelected(packet.get<int8_t>());
+			npc->setSelected(packet.get<uint8_t>());
 		}
 	}
 	else if (type == NPCDialogs::style) {
 		npc->setState(npc->getState() + 1);
 		if (what == 1) {
-			npc->setSelected(packet.get<int16_t>());
+			npc->setSelected(packet.get<uint8_t>());
 		}
 		else  {
 			npc->end();
